requirements.txt
ccxt>=4.0.0
websockets>=12.0
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
loguru>=0.7.0
